Beginning support for server-side connection
This is needed to bypass the parasite in CRIU. It is not trivial to register
a QP in one process and register the memory window in another process, which
is what is required in CRIU since everything is in userspace. Instead, we
are implementing the memory region and QP in kernel space.